+++ examples/RPC/client.php
@@ -14,7 +14,7 @@ use Rubix\Server\Commands\ServerStatus;
 const SHARED_SECRET = '2e2c47bbda4e531c585d796c0c8a4ac9';
 
 $client = new RPCClient('127.0.0.1', 8888, false, [
-    'Authorization' => SHARED_SECRET,
+    'Authorization' => 'Basic ' . base64_encode('test:secret'),
 ]);
 
 $dataset = new Unlabeled([