+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 - 0.2.1-beta
     - Support PSR-15 Server Request Handler interface
+    - Implemented HTTP Basic Auth middleware
 
 - 0.2.0-beta
     - Update to Rubix ML 0.2.0